disable syslog for uWSGI
"syslog:foo" would enable master-worker mode. In this mode,
the application is loaded in master process. Then it is forked
and requests are processed by worker process(es).

The problem is that in this project, prometheus push thread
is created on master process, but metrics data are collected
by worker processes. The pushing mechanism is working, but
it is not sending any data to Prometheus pushgateway :(